+++ mne_bids/io/tests/test_base.py
@@ -1,15 +1,3 @@
-# -*- coding: utf-8 -*-
-"""Test the MNE BIDS converter.
-
-For each supported file format, implement a test.
-"""
-# Authors: Mainak Jas <mainak.jas@telecom-paristech.fr>
-#          Teon L Brooks <teon.brooks@gmail.com>
-#          Chris Holdgraf <choldgraf@berkeley.edu>
-#          Stefan Appelhoff <stefan.appelhoff@mailbox.org>
-#          Matt Sanderson <matt.sanderson@mq.edu.au>
-#
-# License: BSD (3-clause)
 
 import os
 import os.path as op
